+++ CHANGELOG.md
@@ -10,8 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
-- style of radio button. Was ot visible in IE, Chrome, Safari
-- 
+-
+
+## [0.7.0](https://github.com/udigeri/inout/compare/v0.6.0..v0.7.0) (2021-04-12)
+
+### Added
+
+- icons for payment types are loaded from PGS imageUrl
+- support for selection colored or B/W icons
+
 ## [0.6.0](https://github.com/udigeri/inout/compare/v0.5.0..v0.6.0) (2021-04-12)
 
 ### Added